src/main/java/com/biperf/core/domain/enums/PromotionJobPositionType.java
/*
 * (c) 2005 BI, Inc.  All rights reserved.
 * $Source: /usr/local/ndscvsroot/products/beacon/src/java/com/biperf/core/domain/enums/PromotionJobPositionType.java,v $
 */

package com.biperf.core.domain.enums;

import java.util.List;

/*
 * PromotionJobPositionType <p> <b>Change History:</b><br> <table border="1"> <tr> <th>Author</th>
 * <th>Date</th> <th>Version</th> <th>Comments</th> </tr> <tr> <td>OPI Admin</td> <td>Sep
 * 1, 2005</td> <td>1.0</td> <td>created</td> </tr> </table> </p>
 * 
 *
 */

public class PromotionJobPositionType extends PickListItem
{

  /**
   * Asset name used in Content Manager
   */
  private static final String PICKLIST_ASSET = "picklist.promotion.job.position.type";

  /**
   * This constructor is protected - only the PickListFactory class creates these instances.
   */
  protected PromotionJobPositionType()
  {
    super();
  }

  /**
   * Get the pick list from content manager.
   * 
   * @return List
   */
  public static List getList()
  {
    return getPickListFactory().getPickList( PromotionJobPositionType.class );
  }

  /**
   * Returns null if the code is null or invalid for this list.
   * 
   * @param code
   * @return PickListItem
   */
  public static PromotionJobPositionType lookup( String code )
  {
    return (PromotionJobPositionType)getPickListFactory().getPickListItem( PromotionJobPositionType.class, code );
  }

  /**
   * Returns null if the defaultItem is not defined - or invalid
   * 
   * @return default PickListItem
   */
  // public static PromotionJobPositionType getDefaultItem()
  // {
  // return (PromotionJobPositionType)getPickListFactory().getDefaultPickListItem(
  // PromotionJobPositionType.class );
  // }
  /**
   * Overridden from
   * 
   * @see com.biperf.core.domain.enums.PickListItem#getPickListAssetCode()
   * @return PICKLIST_ASSET
   */
  public String getPickListAssetCode()
  {
    return PICKLIST_ASSET;
  }

}
